backend/requirements.txt
asgiref==3.10.0
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
invoke==2.2.1
paramiko==4.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
PyNaCl==1.6.0
python-dotenv==1.1.1
sqlparse==0.5.3
